% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmaster.R
\name{archR_to_seurat}
\alias{archR_to_seurat}
\title{archr matrix to seurat object}
\usage{
archR_to_seurat(proj, matrix, binarize, archr_rd)
}
\arguments{
\item{proj}{}

\item{matrix}{}

\item{binarize}{}

\item{archr_rd}{}
}
\value{
seurat object with specificed umap coordinates
}
\description{
archr matrix to seurat object
}
\examples{
seurat<-archR_to_serat(proj, "GeneScoreMatrix", binarize = F, archr_rd = "ATAC_UMAP")


}
