% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{calculate_fit_metrics}
\alias{calculate_fit_metrics}
\title{Calculate goodness of fit metrics for deconvolution}
\usage{
calculate_fit_metrics(observed, predicted, gene_weights)
}
\arguments{
\item{observed}{Matrix of observed counts (genes × samples)}

\item{predicted}{Matrix of predicted counts (genes × samples)}
}
\description{
Calculate goodness of fit metrics for deconvolution
}
\keyword{internal}
