% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{cluster_specific}
\alias{cluster_specific}
\title{Find Cluster specific genes quickly}
\usage{
cluster_specific(
  cds,
  column = "cluster",
  verbose = T,
  threads = 1,
  run_glm = F,
  min_expression = 0.2,
  top_n = 500,
  symbol_id = "gene_id"
)
}
\arguments{
\item{cds}{cell_data_set object with a reduced dimension matrix (currently LSI supported) as specified in
reduced_dim argument and a model used to create a low dimensional embedding}

\item{column}{a SummarizedExperiment type object (cell_data_set currently supported) to be projected using 
the models contained in the projector}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of this function. Can be set to FALSE for a cleaner output.}

\item{threads}{The number of threads used for parallel execution}
}
\description{
Find Cluster specific genes quickly
}
