% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{common_variant_genes}
\alias{common_variant_genes}
\title{Common Variant Genes}
\usage{
common_variant_genes(
  cds1,
  cds2,
  top_n = 2000,
  logmean_ul = 2,
  logmean_ll = -6,
  row_data_column = "gene_short_name",
  unique_data_column = "id",
  verbose = T,
  plot = F
)
}
\arguments{
\item{cds1}{The first cell data set.}

\item{cds2}{The second cell data set.}

\item{top_n}{The number of top genes to consider. Default is 2000.}

\item{logmean_ul}{The upper limit for mean expression.}

\item{logmean_ll}{The lower limit for mean expression.}

\item{row_data_column}{The column in the feature data corresponding to the gene symbol.}

\item{unique_data_column}{The column in the feature data corresponding to the unique id.}

\item{verbose}{Boolean indicating whether to display verbose output.}

\item{plot}{Boolean indicating whether to plot the results.}
}
\value{
A vector of similarly variant genes.
}
\description{
This function finds common variant genes between two cell data sets.
}
