% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{computeSparseRowVariances}
\alias{computeSparseRowVariances}
\title{Compute sparse row variance}
\usage{
computeSparseRowVariances(j, val, rm, n)
}
\description{
Compute sparse row variance
}
\keyword{internal}
