% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion matrix}
\usage{
confusion_matrix(pred, gt, cols = NULL)
}
\arguments{
\item{pred}{factor of predictions}

\item{gt}{factor of ground truth}

\item{cols}{named vector of colors}
}
\value{
a confusion matrix plot
}
\description{
This function will generate a confusion matrix between two factors; pred (short for prediction) and gt (short for ground truth).  One
may optionally supply a named vector of colors to annotate the row and column legends.
}
