% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmastR.R
\name{demo}
\alias{demo}
\title{demo}
\usage{
demo(
  perc = 80,
  FUNC = c("naive_bayes", "neural_network", "softmax_regression", "deep_belief_nn",
    "logistic_regression", "bagging", "perceptron")
)
}
\arguments{
\item{perc}{percent to holdout}

\item{FUNC}{algorithm to use}
}
\description{
run classification demo
}
