% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tenx.R
\name{files3_prep}
\alias{files3_prep}
\title{Files3 prep - find cellranger data files (matrix.mtx, barcodes.tsv, genes.tsv)}
\usage{
files3_prep(folder)
}
\arguments{
\item{folders}{A vector of cellranger folders (full folder name)}
}
\description{
Files3 prep - find cellranger data files (matrix.mtx, barcodes.tsv, genes.tsv)
}
