% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmastR.R
\name{franken}
\alias{franken}
\title{Franken}
\usage{
franken(
  cds,
  rowdata_col = "gene_short_name",
  from_species = "mm",
  to_species = "hs",
  trim = T
)
}
\arguments{
\item{cds}{cell_data_set}

\item{rowdata_col}{rowData column to lookup}

\item{from_species}{currently only mouse ("mm") and human ("hs") symbols supported}
}
\value{
a cell_data_set object
}
\description{
Will prepare monocle3 objects for use across species
}
