% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmastR.R
\name{franken_helper}
\alias{franken_helper}
\title{franken_helper}
\usage{
franken_helper(
  x,
  from_dataset = "mmusculus_gene_ensembl",
  to_dataset = "hsapiens_gene_ensembl",
  from_type = "mgi_symbol",
  to_type = "hgnc_symbol"
)
}
\arguments{
\item{x}{}

\item{from_dataset}{}

\item{to_dataset}{}

\item{from_type}{}

\item{to_type}{}
}
\description{
franken_helper
}
