% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_features.R
\name{get_selected_features}
\alias{get_selected_features}
\title{Get Selected Features for Ordering}
\usage{
get_selected_features(cds, gene_column = "id")
}
\arguments{
\item{cds}{An object of class \code{cell_data_set} (Monocle3) or \code{Seurat}.}

\item{gene_column}{Character. The name of the column in the dispersion table containing
the gene identifiers you wish to retrieve. Default is "id".}
}
\value{
A character vector containing the identifiers of the selected features.
}
\description{
This function extracts the features marked as \code{use_for_ordering} within the
object's internal dispersion metadata. It supports both Monocle3 (\code{cell_data_set})
and Seurat objects, provided the dispersion data is stored in the expected slots.
}
\details{
Retrieves the list of features (genes) currently selected for trajectory ordering or
downstream analysis from a single-cell object.
}
\examples{
\dontrun{
  # Retrieve IDs of ordering genes
  ordering_genes <- get_selected_features(cds)

  # Retrieve common names if stored in a "gene_short_name" column
  ordering_genes_names <- get_selected_features(cds, gene_column = "gene_short_name")
}
}
