% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmaster.R
\name{humanize}
\alias{humanize}
\title{humanize}
\usage{
humanize(seurat)
}
\arguments{
\item{seurat}{}
}
\value{
seurat object with humanized genes
}
\description{
This functin assumes that you seurat object has mouse genes and will return a new seurat object with 'humanized' genes.
You will have to re-embed ie. NormalizeData() ScaleData() FindVariableFeatures() RunPCA() RunUMAP() after this!!!
}
\examples{
human_seurat<-humanize(mouse_seurat)

}
