% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{infer_from_model}
\alias{infer_from_model}
\title{infer from saved model}
\usage{
infer_from_model(
  model_path,
  query,
  num_classes,
  num_features,
  verbose,
  batch_size
)
}
\description{
infer from saved model
}
\keyword{internal}
