% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmaster.R
\name{log_reg_matrix}
\alias{log_reg_matrix}
\title{Single Cell Logistic Regression Matrix}
\usage{
log_reg_matrix(trainDat, testDat, trainClass, testClass, downsample)
}
\arguments{
\item{trainDat}{}

\item{testDat}{}

\item{trainClass}{}

\item{testClass}{}

\item{downsample}{}
}
\value{

}
\description{
Single Cell Logistic Regression Matrix
}
