% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmaster.R
\name{monocle3_to_seurat}
\alias{monocle3_to_seurat}
\title{Monocle3 to Seurat}
\usage{
monocle3_to_seurat(
  cds,
  seu_rd = "umap",
  mon_rd = "UMAP",
  assay_name = "RNA",
  row.names = "gene_short_name",
  normalize = T
)
}
\arguments{
\item{seu_rd}{Reduced dimname for seurat ('i.e. UMAP')}

\item{mon_rd}{Reduced dimname for monocle3 ('i.e. UMAP')}

\item{assay_name}{Name of data slot ('i.e. RNA')}

\item{row.names}{rowData column to use as rownames for Seurat object}

\item{seu}{Seurat Object}
}
\value{
a cell_data_set object
}
\description{
Conver Monocle3 cell data set to a Seurat object.  For a variety of reasons, the recommendations are to use this funciont
only to generate skeleton Seurat objects that can be used for plotting and not much else.  The resulting object will not contain PCA reducitons or 
nearest neighbor graphs.
}
