% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmastR.R
\name{multinomialFitCV}
\alias{multinomialFitCV}
\title{multinomialFitCV}
\usage{
multinomialFitCV(x, y, nParallel = detectCores(), ...)
}
\arguments{
\item{x}{}

\item{y}{}

\item{nParallel}{}

\item{...}{}
}
\description{
multinomialFitCV
}
