% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_data.R
\name{plot_projection}
\alias{plot_projection}
\title{Plot a projection}
\usage{
plot_projection(
  projection,
  projector,
  projectee,
  projector_col = NULL,
  projectee_col = NULL
)
}
\arguments{
\item{projection}{projection object created by project_data function}

\item{projector}{cell_data_set object with a reduced dimension matrix (currently LSI supported) as specified in
reduced_dim argument and a model used to create a low dimensional embedding}

\item{projectee}{a SummarizedExperiment type object (cell_data_set currently supported) to be projected using 
the models contained in the projector}

\item{projector_col}{column name of projector to color cells by using nearest neighbor}

\item{projectee_col}{column name of projectee to color cells by}

\item{colors}{color palatte of projectee_col cells}
}
\description{
This function plots a projection object created using the project_data function
}
