% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_training_data}
\alias{plot_training_data}
\title{Training data plot}
\usage{
plot_training_data(output_list)
}
\arguments{
\item{output_list}{a list returned from running viewmastR using the return_type = "list" parameter.}
}
\value{
a plot of training data
}
\description{
This function will display a plot of data generated during viewmastR training
}
