% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{print_metrics_summary}
\alias{print_metrics_summary}
\title{Print summary of deconvolution metrics}
\usage{
print_metrics_summary(result)
}
\description{
Print summary of deconvolution metrics
}
\keyword{internal}
