% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{process_learning_obj_ann}
\alias{process_learning_obj_ann}
\title{Process Robj learning objects for ANN}
\usage{
process_learning_obj_ann(
  train,
  test,
  query,
  labels,
  hidden_size,
  learning_rate,
  num_epochs,
  directory,
  verbose,
  backend
)
}
\description{
Process Robj learning objects for ANN
}
\keyword{internal}
