% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{process_learning_obj_mlr}
\alias{process_learning_obj_mlr}
\title{Process Robj learning objects for MLR}
\usage{
process_learning_obj_mlr(
  train,
  test,
  query,
  labels,
  learning_rate,
  num_epochs,
  directory,
  verbose,
  backend
)
}
\description{
Process Robj learning objects for MLR
}
\keyword{internal}
