% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{process_learning_obj_nb}
\alias{process_learning_obj_nb}
\title{Run test nb training}
\usage{
process_learning_obj_nb(train, test, query)
}
\description{
Run test nb training
}
\keyword{MLR}
\keyword{Process}
\keyword{Robj}
\keyword{for}
\keyword{internal}
\keyword{learning}
\keyword{objects}
