% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmastR.R
\name{pseudo_singlets}
\alias{pseudo_singlets}
\title{Pseudo-singlets}
\usage{
pseudo_singlets(
  se,
  sc_cds,
  assay_name = "logcounts",
  logtransformed = T,
  selected_genes = NULL,
  ncells_per_col = 50,
  threads = detectCores()
)
}
\arguments{
\item{se}{Summarized Experiment}

\item{cds}{reference cds}
}
\value{
a cell_data_set object or a list of items if unfiltered data is returned (see unfiltered)
}
\description{
ip
}
