% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{setup_training}
\alias{setup_training}
\title{Setup training datasets}
\usage{
setup_training(
  query_cds,
  ref_cds,
  ref_celldata_col,
  norm_method = c("log", "binary", "size_only", "none"),
  selected_genes = NULL,
  train_frac = 0.8,
  tf_idf = F,
  scale = F,
  LSImethod = 1,
  verbose = T,
  addbias = F,
  use_sparse = F,
  return_type = c("list", "matrix", "S4obj"),
  ...
)
}
\arguments{
\item{query_cds}{cds to query}

\item{ref_cds}{reference cds}
}
\value{
various forms of training test data and query
}
\description{
see viewmastR
}
\keyword{internal}
