% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmastR.R
\name{seurat_to_monocle3}
\alias{seurat_to_monocle3}
\title{Seurat to Monocle3}
\usage{
seurat_to_monocle3(seu, seu_rd = "umap", mon_rd = "UMAP", assay_name = "RNA")
}
\arguments{
\item{seu}{Seurat Object}

\item{seu_rd}{Reduced dimname for seurat ('i.e. UMAP')}

\item{mon_rd}{Reduced dimname for monocle3 ('i.e. UMAP')}

\item{assay_name}{Name of data slot ('i.e. RNA')}
}
\value{
a cell_data_set object
}
\description{
Conver Seurat to Monocle3
}
