% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{sfc}
\alias{sfc}
\title{Make colors}
\usage{
sfc(n, scramble = F)
}
\arguments{
\item{n}{number of colors}

\item{scramble}{whether to jumble the colors}
}
\description{
This function creates a color palatte
}
