% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/viewmaster.R
\docType{package}
\name{viewmaster}
\alias{viewmaster}
\title{viewmaster}
\usage{
viewmaster(
  query_cds,
  ref_cds,
  ref_celldata_col = "celltype",
  query_celldata_col = NULL,
  FUNC = c("naive_bayes", "neural_network", "bagging", "softmax_regression",
    "logistic_regression", "deep_belief_nn"),
  selected_genes = NULL,
  train_frac = 0.8,
  verbose = T
)
}
\arguments{
\item{query_cds}{cds to query}

\item{ref_cds}{reference cds}
}
\value{
a cell_data_set object or a list of items if unfiltered data is returned (see unfiltered)
}
\description{
see single cell datasets through different lenses

ip
}
\author{
scfurl
}
