% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{viewmastR-package}
\alias{viewmastR-package}
\title{viewmastR: Automated single-cell genomic cell type assignment}
\description{
viewmastR uses machine learning implemented in Rust to perform automated cell type classification for single-cell genomic data. Currently viewmastR is authored to work with scRNAseq data, but more features are coming.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://furlan-lab.github.io/viewmastR/}
}

}
\author{
\strong{Maintainer}: Scott Furlan \email{scottfurlan@gmail.com} (\href{https://orcid.org/0000-0003-3007-9840}{ORCID})

}
