% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{viewmastR}
\alias{viewmastR}
\title{Running viewmastR using the new Rust implementation}
\usage{
viewmastR(
  query_cds,
  ref_cds,
  ref_celldata_col,
  query_celldata_col = NULL,
  FUNC = c("softmax_regression"),
  norm_method = c("log", "binary", "size_only", "none"),
  selected_genes = NULL,
  train_frac = 0.8,
  tf_idf = F,
  scale = F,
  hidden_layers = c(500, 100),
  learning_rate = 0.001,
  batch_size = 100,
  max_epochs = 10,
  max_error = 0.5,
  lambda = 1,
  iterations = 1000,
  LSImethod = 1,
  verbose = T,
  device = 0,
  threshold = NULL,
  keras_model = NULL,
  dir = "/tmp/sc_local",
  return_type = c("object", "list"),
  ...
)
}
\arguments{
\item{query_cds}{cds to query}

\item{ref_cds}{reference cds}
}
\value{
various forms of training test data and query
}
\description{
see viewmastR
}
