% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{viewmastrPal}
\alias{viewmastrPal}
\title{List of color palettes that can be used in plots}
\format{
An object of class \code{list} of length 33.
}
\usage{
viewmastrPal
}
\description{
A collection of some original and some borrowed color palettes to provide appealing color aesthetics for plots in viewmastR
}
\keyword{datasets}
