% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{writeMMgz}
\alias{writeMMgz}
\title{Write a compressed MatrixMarket file}
\usage{
writeMMgz(x, file)
}
\arguments{
\item{x}{A sparse matrix (typically a \code{dgCMatrix} or \code{ngCMatrix} object).}

\item{file}{A character string specifying the output file name, which will be compressed into \code{.gz} format.}
}
\value{
This function does not return a value. It writes a file as a side effect.
}
\description{
This function writes a sparse matrix in the MatrixMarket format to a compressed \code{.gz} file.
The function handles both real and integer matrix types.
}
\details{
This function writes the matrix in the MatrixMarket coordinate format.
It first writes the header indicating the matrix type and size, and then appends the matrix data.
If the matrix is an \code{ngCMatrix}, it is treated as an integer matrix, otherwise as a real matrix.
The function compresses the output into a \code{.gz} file.
}
\examples{
\dontrun{
library(Matrix)
m <- Matrix(c(0, 1, 0, 2), 2, 2, sparse = TRUE)
writeMMgz(m, "matrix.mtx.gz")
}
}
