% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmaster.R
\name{archR_to_seurat}
\alias{archR_to_seurat}
\title{archR_to_seurat}
\usage{
archR_to_seurat(proj, matrix, binarize, archr_rd)
}
\arguments{
\item{proj}{}

\item{matrix}{}

\item{binarize}{}

\item{archr_rd}{}
}
\description{
archR_to_seurat
}
