% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmastR.R
\name{common_variant_genes}
\alias{common_variant_genes}
\title{Common Variant Genes}
\usage{
common_variant_genes(
  cds1,
  cds2,
  top_n = 2000,
  logmean_ul = 2,
  logmean_ll = -6,
  row_data_column = "gene_short_name",
  unique_data_column = "id",
  verbose = T
)
}
\arguments{
\item{cds1}{cds}

\item{cds2}{}
}
\value{
a vector of similarly variant genes
}
\description{
Find common variant genes between two cds objects
}
