% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmaster.R
\name{humanize}
\alias{humanize}
\title{humanize}
\usage{
humanize(seurat, seu_rd, assay)
}
\arguments{
\item{seurat}{}
}
\value{
seurat object with humanized genes
}
\description{
This function assumes that you seurat object has mouse genes and will return a new seurat object with 'humanized' genes.
You will lose variable features after this, the original umap coordinates are maintained
}
\examples{
human_seurat<-humanize(mouse_seurat)

}
