% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmaster.R
\name{log_reg_matrix}
\alias{log_reg_matrix}
\title{Single Cell Logistic Regression Matrix}
\usage{
log_reg_matrix(trainDat, testDat, trainClass, testClass, downsample = 200)
}
\arguments{
\item{trainDat}{training seurat object, assumes the negative control is in it already}

\item{testDat}{testing seurat object}

\item{trainClass}{metadata column to learn}

\item{testClass}{metadata column to test}

\item{downsample}{number of randomly downsampled cells per metadata column in trainDat}
}
\description{
Single Cell Logistic Regression Matrix
}
