% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{process_learning_obj_nb}
\alias{process_learning_obj_nb}
\title{Process Robj learning objects for MLR}
\usage{
process_learning_obj_nb(train, test, query)
}
\description{
Process Robj learning objects for MLR
}
\keyword{internal}
