% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{setup_training}
\alias{setup_training}
\title{Setup training datasets}
\usage{
setup_training(
  query_cds,
  ref_cds,
  ref_celldata_col,
  norm_method = c("log", "binary", "size_only", "none"),
  selected_genes = NULL,
  train_frac = 0.8,
  tf_idf = F,
  scale = F,
  LSImethod = 1,
  verbose = T,
  addbias = F,
  return_type = c("list", "matrix", "S4obj"),
  ...
)
}
\arguments{
\item{query_cds}{A cell data set (cds) to query.}

\item{ref_cds}{A reference cell data set (cds).}

\item{ref_celldata_col}{The column in the reference cell data set containing cell data.}

\item{norm_method}{The normalization method to use. Options are "log", "binary", "size_only", or "none".}

\item{selected_genes}{A vector of pre-selected genes for analysis.}

\item{train_frac}{The fraction of data to use for training.}

\item{tf_idf}{Boolean indicating whether to perform TF-IDF transformation.}

\item{scale}{Boolean indicating whether to scale the data.}

\item{LSImethod}{The method for Latent Semantic Indexing.}

\item{verbose}{Boolean indicating whether to display verbose output.}

\item{addbias}{Boolean indicating whether to add bias.}

\item{return_type}{The type of output to return. Options are "list", "matrix", or "S4obj".}

\item{...}{Additional arguments.}
}
\value{
A list, matrix, or S4 object containing the training datasets.
}
\description{
This function sets up training datasets for use in machine learning models.
}
\keyword{internal}
