% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/viewmastR.R
\docType{package}
\name{viewmastR}
\alias{viewmastR}
\title{viewmastR}
\usage{
viewmastR(
  query_cds,
  ref_cds,
  ref_celldata_col,
  query_celldata_col = NULL,
  FUNC = c("naive_bayes", "neural_network", "bagging", "softmax_regression",
    "logistic_regression", "deep_belief_nn", "perceptron", "keras_nn", "xgboost",
    "lasso"),
  selected_genes = NULL,
  train_frac = 0.8,
  tf_idf = F,
  hidden_layers = c(500, 100),
  learning_rate = 0.5,
  batch_size = 100,
  max_epochs = 250,
  max_error = 0.5,
  lambda = 1,
  iterations = 1000,
  LSImethod = 1,
  verbose = T,
  device = 0,
  threshold = NULL,
  keras_model = NULL,
  ...
)
}
\arguments{
\item{query_cds}{cds to query}

\item{ref_cds}{reference cds}
}
\value{
a cell_data_set object or a list of items if unfiltered data is returned (see unfiltered)
}
\description{
see single cell datasets through different lenses

in progress
}
\author{
scfurl
}
