% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{viewmastR_infer}
\alias{viewmastR_infer}
\title{Function to infer cell labels using a trained model}
\usage{
viewmastR_infer(
  query_cds,
  model_path,
  vg,
  query_celldata_col = "viewmastR_inferred",
  labels = NULL,
  verbose = TRUE,
  return_probs = FALSE,
  return_type = c("object", "list"),
  chunks = 1,
  workers = 1
)
}
\arguments{
\item{query_cds}{Seurat or cell_data_set object}

\item{model_path}{character path to the trained model file}

\item{vg}{character vector of features used for inference}

\item{query_celldata_col}{character column name to store inferred cell labels}

\item{labels}{character vector of labels (optional)}

\item{verbose}{logical, show messages}

\item{return_probs}{logical, return class probabilities}

\item{return_type}{one of "object" or "list"}

\item{chunks}{integer number of chunks to split the data into for parallelization}

\item{workers}{integer number of parallel workers}
}
\description{
Function to infer cell labels using a trained model
}
